Fixed t-stat error & added error handle to mass regression function
The t-statistic for the no BMG table was not generating correctly as the transformation from list to table had been omitted earlier in the code.
The error handle for the mass regression function checked the number of rows in the table, but not the number of rows without an NA. As such, a table that was only NAs was slipping through.
The full Fama-French data was added.