Added script to create BMG based on carve-outs of index
Added R/custom_carveouts_bmg.R which takes in a list of green stocks and creates a carve-out of them from the primary index. The remaining stocks then are considered brown stocks. This then allows a custom brown minus green index to be created and saved. The script is flexible enough that the "brown" index can also be its own subset of the primary index and not the remainder of the stocks
Also added some extra output from the clean_bmg_validation_script so that the loadings for all the stocks are saved